Break js functions out into separate files and directories
